Java endpoint to fetch URL (JSON or CSV). second commit upon team review
