+++ src/main/java/com/iscte/engsoft/grupob/CalendarApp/model/CalendarFormat.java
@@ -1,0 +1,6 @@
+package com.iscte.engsoft.grupob.CalendarApp.model;
+
+public enum CalendarFormat {
+        CSV, JSON
+
+}
